Using java WebSocket client.
The WebSocket support in QtWebKit(Qt 4.8.2) is far behind the upstream:
http://trac.webkit.org/changeset/89669
http://trac.webkit.org/changeset/110905

To work with Socko WebSocket server, a newer protocol need to be
support. Using WebSocket client library from http://java-websocket.org/, with
the help of QtWebKit bridge, a simple api is exposed the javascript. Because
the limitation of QtWebKit bridge, all callback functions in server.js are
changed to use 'url' as first argument instead of original 'socket'.

The code works under simple test, but may fail in more complex scenario.